+++ src/jtl/Connector/Shopware/Mapper/CustomerOrder.php
@@ -278,7 +278,7 @@ class CustomerOrder extends DataMapper
 
         $shopMapper = Mmc::getMapper('Shop');
         $shops = $shopMapper->findByLocale($localeSW->getLocale());
-        if ($shops === null) {
+        if ($shops === null || (is_array($shops) && count($shops) == 0)) {
             throw new \Exception(sprintf('Shop with language iso (%s) and locale (%s) not found', $customerOrder->getLanguageISO(), $localeSW->getLocale()));
         }