+++ src/jtl/Connector/Shopware/Mapper/Image.php
@@ -6,15 +6,19 @@
 
 namespace jtl\Connector\Shopware\Mapper;
 
+use jtl\Connector\Core\IO\Path;
 use jtl\Connector\Core\Logger\Logger;
 use jtl\Connector\Core\Utilities\Seo;
 use \jtl\Connector\Drawing\ImageRelationType;
 use jtl\Connector\Formatter\ExceptionFormatter;
+use jtl\Connector\Linker\IdentityLinker;
 use \jtl\Connector\Model\Image as ImageModel;
 use \jtl\Connector\Shopware\Model\Image as ImageConModel;
 use \jtl\Connector\Model\Identity;
 use \Shopware\Models\Media\Media as MediaSW;
 use \Shopware\Models\Article\Image as ArticleImageSW;
+use \Shopware\Models\Article\Image\Mapping as MappingSW;
+use \Shopware\Models\Article\Image\Rule as RuleSW;
 use \jtl\Connector\Shopware\Utilities\Mmc;
 use \jtl\Connector\Shopware\Utilities\IdConcatenator;
 use \Shopware\Models\Article\Detail as DetailSW;
@@ -28,6 +32,29 @@ class Image extends DataMapper
         return $this->Manager()->getRepository('Shopware\Models\Media\Media')->find((int) $id);
     }
 
+    public function findBy(array $kv)
+    {
+        return $this->Manager()->getRepository('Shopware\Models\Media\Media')->findOneBy($kv);
+    }
+
+    public function findArticleImage($id)
+    {
+        try {
+            return $this->Manager()->createQueryBuilder()
+                ->select(
+                    'image',
+                    'media'
+                )
+                ->from('Shopware\Models\Article\Image', 'image')
+                ->leftJoin('image.media', 'media')
+                ->where('image.id = :id')
+                ->setParameter('id', $id)
+                ->getQuery()->getOneOrNullResult();
+        } catch (\Exception $e) {
+            return null;
+        }
+    }
+
     public function findAll($limit = null, $count = false, $relationType = null)
     {
         $rsm = new \Doctrine\ORM\Query\ResultSetMapping();
@@ -37,6 +64,30 @@ class Image extends DataMapper
 
         switch ($relationType) {
             case ImageRelationType::TYPE_PRODUCT:
+                return Shopware()->Db()->fetchAssoc(
+                    'SELECT i.id as cId, if (d.id > 0, d.id, a.main_detail_id) as detailId, i.*, m.path
+                      FROM s_articles_img i
+                      LEFT JOIN s_articles_img c ON c.parent_id = i.id
+                      LEFT JOIN s_articles a ON a.id = i.articleID
+                      LEFT JOIN s_articles_details d ON d.articleID = a.id
+                          AND d.kind = 0
+                      LEFT JOIN jtl_connector_link_product_image l ON l.id = i.id
+                      JOIN s_media m ON m.id = i.media_id
+                      WHERE i.articleID IS NOT NULL
+                          AND c.id IS NULL
+                          AND l.host_id IS NULL
+                      UNION
+                      SELECT i.id as cId, i.article_detail_id as detailId, p.*, m.path
+                      FROM s_articles_img i
+                      JOIN s_articles_img p ON i.parent_id = p.id
+                      LEFT JOIN jtl_connector_link_product_image l ON l.id = i.id
+                      JOIN s_media m ON m.id = p.media_id
+                      WHERE i.articleID IS NULL
+                          AND l.host_id IS NULL
+                      LIMIT ' . intval($limit)
+                );
+
+                /*
                 return $this->Manager()->createQueryBuilder()
                     ->select(
                         'image',
@@ -55,6 +106,7 @@ class Image extends DataMapper
                     ->setMaxResults($limit)
                     ->where('linker.hostId IS NULL')
                     ->getQuery()->getResult(\Doctrine\ORM\AbstractQuery::HYDRATE_ARRAY);
+                */
                 break;
             case ImageRelationType::TYPE_CATEGORY:
                 $query = Shopware()->Models()->createNativeQuery(
@@ -92,11 +144,37 @@ class Image extends DataMapper
         $count = 0;
         switch ($relationType) {
             case ImageRelationType::TYPE_PRODUCT:
+                $counts = Shopware()->Db()->fetchAssoc(
+                    'SELECT count(*) as count
+                      FROM s_articles_img i
+                      LEFT JOIN s_articles_img c ON c.parent_id = i.id
+                      LEFT JOIN s_articles a ON a.id = i.articleID
+                      LEFT JOIN jtl_connector_link_product_image l ON l.id = i.id
+                      JOIN s_media m ON m.id = i.media_id
+                      WHERE i.articleID IS NOT NULL
+                          AND c.id IS NULL
+                          AND l.host_id IS NULL
+                      UNION
+                      SELECT count(*) as count
+                      FROM s_articles_img i
+                      JOIN s_articles_img p ON i.parent_id = p.id
+                      LEFT JOIN jtl_connector_link_product_image l ON l.id = i.id
+                      JOIN s_media m ON m.id = p.media_id
+                      WHERE i.articleID IS NULL
+                          AND l.host_id IS NULL'
+                );
+
+                foreach ($counts as $c) {
+                    $count += (int) $c['count'];
+                }
+
+                /*
                 $query = Shopware()->Models()->createNativeQuery(
                     'SELECT count(*) as count 
                     FROM s_articles_img a
                     LEFT JOIN jtl_connector_link_product_image p ON p.id = a.id
                     WHERE p.host_id IS NULL', $rsm);
+                */
                 break;
             case ImageRelationType::TYPE_CATEGORY:
                 $query = Shopware()->Models()->createNativeQuery(
@@ -156,11 +234,37 @@ class Image extends DataMapper
 
             $this->flush();
 
+            // Save product image variation mappings, if product is a child
+            if ($imageSW !== null && $imageSW instanceof ArticleImageSW && $imageSW->getParent() !== null
+                && $image->getRelationType() === ImageRelationType::TYPE_PRODUCT) {
+
+                // Save mapping and rule
+                $this->saveImageMapping($imageSW->getParent());
+                $this->flush();
+            }
+
             $manager = Shopware()->Container()->get('thumbnail_manager');
             $manager->createMediaThumbnail($mediaSW, array(), true);
 
+            $endpoint = ImageConModel::generateId($image->getRelationType(), $imageSW->getId(), $mediaSW->getId());
+            if (strlen($image->getId()->getEndpoint()) > 0 && $image->getId()->getHost() > 0
+                && $endpoint !== $image->getId()->getEndpoint()) {
+
+                Application()->getConnector()->getPrimaryKeyMapper()->delete(
+                    $image->getId()->getEndpoint(),
+                    $image->getId()->getHost(),
+                    IdentityLinker::TYPE_IMAGE
+                );
+
+                Application()->getConnector()->getPrimaryKeyMapper()->save(
+                    $endpoint,
+                    $image->getId()->getHost(),
+                    IdentityLinker::TYPE_IMAGE
+                );
+            }
+
             // Result
-            $result->setId(new Identity(ImageConModel::generateId($image->getRelationType(), $imageSW->getId(), $mediaSW->getId()), $image->getId()->getHost()))
+            $result->setId(new Identity($endpoint, $image->getId()->getHost()))
                 ->setForeignKey(new Identity($image->getForeignKey()->getEndpoint(), $image->getForeignKey()->getHost()))
                 ->setRelationType($image->getRelationType())
                 ->setFilename(sprintf('http://%s%s/%s', Shopware()->Shop()->getHost(), Shopware()->Shop()->getBaseUrl(), $mediaSW->getPath()));
@@ -177,6 +281,7 @@ class Image extends DataMapper
     {
         list($type, $imageId, $mediaId) = IdConcatenator::unlink($image->getId()->getEndpoint());
 
+        $deleteMedia = true;
         switch ($image->getRelationType()) {
             case ImageRelationType::TYPE_PRODUCT:
                 $foreignId = (strlen($image->getForeignKey()->getEndpoint()) > 0) ? $image->getForeignKey()->getEndpoint() : null;
@@ -188,6 +293,14 @@ class Image extends DataMapper
 
                 // Special delete (masterkill) all images for a single product call
                 if ($image->getSort() == 0 && strlen($image->getId()->getEndpoint()) == 0) {
+                    $mediaResults = Shopware()->Db()->fetchAssoc(
+                        'SELECT i.media_id
+                         FROM s_articles_img i
+                         JOIN s_media m ON m.id = i.media_id
+                         WHERE i.articleID = ' . intval($articleId) . '
+                         GROUP BY i.media_id'
+                    );
+
                     Shopware()->Db()->query(
                         'DELETE i, m, r
                           FROM s_articles_img i
@@ -204,19 +317,51 @@ class Image extends DataMapper
                         ->getQuery()
                         ->execute();
 
+                    $manager = Shopware()->Container()->get('thumbnail_manager');
+                    foreach ($mediaResults as $mediaResult) {
+                        $mediaSW = $this->find($mediaResult['media_id']);
+                        if ($mediaSW !== null) {
+                            $manager->removeMediaThumbnails($mediaSW);
+                            @unlink(sprintf('%s%s', Shopware()->DocPath(), $mediaSW->getPath()));
+                            $this->Manager()->remove($mediaSW);
+                        }
+                    }
+
+                    $this->Manager()->flush();
+
                     return;
                 }
 
                 $imageSW = $this->Manager()->getRepository('Shopware\Models\Article\Image')->find((int) $imageId);
+
                 if ($imageSW !== null) {
-                    $imageChildSW = $this->loadChildImage($detailId, $imageSW->getId());
+                    if ($imageSW->getParent() !== null) {
+                        $isParentRemoved = false;
+
+                        if (!$this->isParentImageInUse($imageSW->getParent()->getId(), $detailId) && !$this->isParentImageInUseByMain($imageSW->getId(), $mediaId)) {
+                            $this->Manager()->remove($imageSW->getParent());
+                            $isParentRemoved = true;
+                        } else {
+                            $deleteMedia = false;
+                        }
+
+                        $this->deleteProductImageMappings($imageSW->getParent()->getId());
+                        if (!$isParentRemoved) {
+                            $this->buildProductImageMappings($imageSW->getParent(), $detailId);
+                        }
+                    }
+
+                    /*
+                    $imageChildSW = $this->loadChildImage($imageSW->getId(), $detailId);
 
                     //Check if product is a configurator and check if the image is in use from another child
                     if ($this->isParentImageInUse($imageSW->getId(), $detailId)) {
+                        $deleteMedia = false;
                         $imageSW = $imageChildSW;
                     } elseif ($imageChildSW !== null) {
                         $this->Manager()->remove($imageChildSW);
                     }
+                    */
 
                     try {
                         $this->Manager()->remove($imageSW);
@@ -254,6 +399,13 @@ class Image extends DataMapper
                 break;
         }
 
+        if ($deleteMedia) {
+            $this->deleteMedia($mediaId);
+        }
+    }
+
+    protected function deleteMedia($mediaId)
+    {
         $mediaSW = $this->Manager()->getRepository('Shopware\Models\Media\Media')->find((int) $mediaId);
         if ($mediaSW !== null) {
             @unlink(sprintf('%s%s', Shopware()->OldPath(), $mediaSW->getPath()));
@@ -324,111 +476,73 @@ class Image extends DataMapper
             throw new \Exception(sprintf('Cannot find product with id (%s)', $articleId));
         }
 
-        // Filename SEO
         list ($uuid, $ext) = explode('.', $image->getFilename());
-        $seo = new Seo();
 
-        $productSeo = sprintf('%s %s %s',
-            $productSW->getName(),
-            $detailSW->getAdditionalText(),
-            $image->getSort()
-        );
-
-        if (strlen($productSeo) > 60) {
-            $pos = strpos($productSeo, ' ', 60);
-            if ($pos === false) {
-                $pos = 60;
-            }
-
-            $productSeo = substr($productSeo, 0, $pos);
-        }
-
-        $filename = sprintf('%s.%s', $seo->create(
-            sprintf('%s %s',
-                $productSeo,
-                $detailSW->getNumber()
-            )
-        ), $ext);
+        // Seo
+        $productSeo = $this->getProductSeoName($productSW, $detailSW, $image);
+        $filename = sprintf('%s.%s', $productSeo, $ext);
 
         if (strlen($filename) > 100) {
             $filename = substr($filename, strlen($filename) - 100, 100);
         }
 
-        $path = sys_get_temp_dir() . DIRECTORY_SEPARATOR . $filename;
+        $path = Path::combine(sys_get_temp_dir(), $filename);
         if (copy($image->getFilename(), $path)) {
             $file = new File($path);
             $image->setFilename($path);
         }
 
-        $mediaSW = $this->getMedia($image, $file);
+        $parentExists = false;
+        $childImageSW = null;
+        if ($productMapper->isChildSW($productSW, $detailSW)) {
+            $parentImageSW = $this->findParentImage($articleId, $image->getFilename());
+            if ($parentImageSW && $parentImageSW !== null) {
+                $imageSW = $parentImageSW;
+                $mediaSW = $parentImageSW->getMedia();
+                $parentExists = true;
+            }
+        }
 
-        // parent image
-        $imageSW = $this->getParentImage($image, $mediaSW, $productSW);
+        if (!$parentExists) {
+            $mediaSW = $this->getMedia($image, $file);
+
+            // parent image
+            $imageSW = $this->getParentImage($image, $mediaSW, $productSW);
+        }
 
         // Varkombi?
-        if ($productSW->getConfiguratorSet() !== null) {
-            if ($imageSW->getId() > 0 && $this->isParentImageInUse($imageSW->getId(), $detailSW->getId())) {
+        if ($productMapper->isChildSW($productSW, $detailSW)) {
+            if (!$parentExists && $imageSW->getId() > 0 && $this->isParentImageInUse($imageSW->getId(), $detailSW->getId())) {
                 // Wenn es noch Kinder gibt die das Vaterbild nutzen, lege ein neues Vaterbild an
                 $mediaSW = $this->getNewMedia($image, $file);
+
+                $this->saveImageMapping($imageSW, $detailSW->getId());
 
                 $imageSW = $this->newParentImage($image, $mediaSW, $productSW);
                 $collection = new \Doctrine\Common\Collections\ArrayCollection;
                 $collection->add($imageSW);
                 $mediaSW->setArticles($collection);
-            } else {
-                $this->copyNewMedia($image, $mediaSW, $file);
             }
-
-            $productMapper = Mmc::getMapper('Product');
 
             // if detail is a child
-            if ($imageSW->getParent() === null && $productMapper->isChildSW($productSW, $detailSW)) {
+            if ($imageSW->getParent() === null) {
                 $childImageSW = $this->getChildImage($image, $mediaSW, $detailSW, $imageSW);
+
                 $this->Manager()->persist($childImageSW);
 
-                // Save mapping and rule
-                $this->saveImageMapping($imageSW, $detailSW);
+                $imageSW = $childImageSW;
             }
         } else {
-            $this->copyNewMedia($image, $mediaSW, $file);
+            $this->copyNewMedia($image, $mediaSW, $file, $parentExists);
         }
 
-        $this->copyNewMedia($image, $mediaSW, $file);
+        $this->copyNewMedia($image, $mediaSW, $file, $parentExists);
     }
 
-    protected function saveImageMapping(ArticleImageSW $imageSW, DetailSW $detailSW)
+    protected function saveImageMapping(ArticleImageSW $parentSW, $detailId = null)
     {
-        $mappingSW = null;
-        if ($imageSW->getId() > 0) {
-            $mappingSW = $this->Manager()->getRepository('Shopware\Models\Article\Image\Mapping')->findOneBy(array('image' => $imageSW->getId()));
-        }
-
-        if ($mappingSW === null) {
-            $mappingSW = new \Shopware\Models\Article\Image\Mapping();
-            $this->Manager()->persist($mappingSW);
-        } else {
-            Shopware()->Db()->delete('s_article_img_mapping_rules', array('mapping_id = ?' => $mappingSW->getId()));
-        }
-
-        $mappingSW->setImage($imageSW);
-
-        $relations = Shopware()->Db()->fetchAssoc(
-            'SELECT * FROM s_article_configurator_option_relations WHERE article_id = ?',
-            array($detailSW->getId())
-        );
-
-        $optionMapper = Mmc::getMapper('ConfiguratorOption');
-        foreach ($relations as $relation) {
-            $optionSW = $optionMapper->find((int) $relation['option_id']);
-
-            if ($optionSW !== null) {
-                $ruleSW = new \Shopware\Models\Article\Image\Rule();
-                $ruleSW->setMapping($mappingSW);
-                $ruleSW->setOption($optionSW);
-
-                $this->Manager()->persist($ruleSW);
-            }
-        }
+        $this->deleteProductImageMappings($parentSW->getId());
+        $this->buildProductImageMappings($parentSW, $detailId);
     }
 
     protected function isParentImageInUse($parentId, $detailId)
@@ -447,7 +561,19 @@ class Image extends DataMapper
         return (is_array($results) && count($results) > 0);
     }
 
-    protected function loadChildImage($detailId, $parentId)
+    protected function isParentImageInUseByMain($childImageId, $mediaId)
+    {
+        $count = Shopware()->Db()->fetchOne(
+            'SELECT count(*) as count
+              FROM jtl_connector_link_product_image l
+              WHERE l.id != ' . intval($childImageId) . '
+                  AND l.media_id = ' . intval($mediaId)
+        );
+
+        return ($count !== null && intval($count) > 0);
+    }
+
+    protected function loadChildImage($parentId, $detailId)
     {
         try {
             return $this->Manager()->createQueryBuilder()
@@ -467,7 +593,14 @@ class Image extends DataMapper
 
     protected function getChildImage(ImageModel $image, MediaSW $mediaSW, DetailSW $detailSW, ArticleImageSW $parentImageSW)
     {
-        $imageSW = $this->loadChildImage($detailSW->getId(), $parentImageSW->getId());
+        $imageId = (strlen($image->getId()->getEndpoint()) > 0) ? $image->getId()->getEndpoint() : null;
+
+        //$imageSW = $this->loadChildImage($parentImageSW->getId(), $detailSW->getId());
+
+        if ($imageId !== null) {
+            list($type, $id, $mediaId) = IdConcatenator::unlink($imageId);
+            $imageSW = $this->Manager()->getRepository('Shopware\Models\Article\Image')->find((int) $id);
+        }
 
         if ($imageSW === null) {
             $imageSW = new ArticleImageSW;
@@ -495,7 +628,11 @@ class Image extends DataMapper
         // Try to load Image
         if ($imageId !== null) {
             list($type, $id, $mediaId) = IdConcatenator::unlink($imageId);
-            $imageSW = $this->Manager()->getRepository('Shopware\Models\Article\Image')->find((int) $id);
+            $imageSW = $this->findArticleImage($id);
+
+            if ($imageSW !== null && $imageSW->getParent() !== null) {
+                $imageSW = $imageSW->getParent();
+            }
         }
 
         // New Image
@@ -517,6 +654,13 @@ class Image extends DataMapper
         if ($imageSW->getParent() === null) {
             $imageSW->setPath($mediaSW->getName());
             $imageSW->setMedia($mediaSW);
+
+            if ($imageSW->getId() > 0) {
+                Shopware()->Db()->query(
+                    'UPDATE s_articles_img SET main = ?, position = ? WHERE id = ?',
+                    [$main, $image->getSort(), $imageSW->getId()]
+                );
+            }
         }
 
         return $imageSW;
@@ -718,9 +862,10 @@ class Image extends DataMapper
         return $mediaSW;
     }
 
-    protected function copyNewMedia(ImageModel $image, MediaSW &$mediaSW, File $file)
+    protected function copyNewMedia(ImageModel $image, MediaSW &$mediaSW, File $file, $parentExists = false)
     {
-        if ($mediaSW->getId() > 0 && file_exists($image->getFilename()) && $this->generadeMD5($mediaSW->getPath()) != md5_file($image->getFilename())) {
+        //if ($mediaSW->getId() > 0 && file_exists($image->getFilename()) && $this->generadeMD5($mediaSW->getPath()) !== md5_file($image->getFilename())) {
+        if (!$parentExists && $mediaSW->getId() > 0 && file_exists($image->getFilename())) {
             $stats = stat($image->getFilename());
             $infos = pathinfo($image->getFilename());
 
@@ -728,6 +873,7 @@ class Image extends DataMapper
             $manager->removeMediaThumbnails($mediaSW);
 
             @unlink(sprintf('%s%s', Shopware()->DocPath(), $mediaSW->getPath()));
+
             $file = $file->move($this->getUploadDir(), $image->getFilename());
             $mediaSW->setFileSize($stats['size'])
                 ->setExtension(strtolower($infos['extension']))
@@ -736,19 +882,142 @@ class Image extends DataMapper
         }
     }
 
-    protected function isChild(ImageModel &$image)
-    {
-        return (strlen($image->getForeignKey()->getEndpoint()) > 0 && strpos($image->getForeignKey()->getEndpoint(), '_') !== false);
-    }
-
-    protected function getUploadDir()
+    protected function getUploadDir($relativeley = false, $stripLastSlash = false)
     {
         // the absolute directory path where uploaded documents should be saved
-        return Shopware()->DocPath('media_' . strtolower(MediaSW::TYPE_IMAGE));
+        $path = Shopware()->DocPath('media_' . strtolower(MediaSW::TYPE_IMAGE));
+
+        if ($relativeley) {
+            $path = str_replace(Shopware()->OldPath(), '', $path);
+        }
+
+        return $stripLastSlash ? substr($path, 0, strrpos($path, DIRECTORY_SEPARATOR)) : $path;
     }
 
     protected function generadeMD5($path)
     {
         return md5_file(sprintf('%s%s', Shopware()->DocPath(), $path));
+    }
+
+    protected function getProductSeoName(ArticleSW $productSW, DetailSW $detailSW, ImageModel $image)
+    {
+        $seo = new Seo();
+
+        $pk = ' ' . $image->getId()->getHost();
+        if ($productSW->getConfiguratorSet() !== null && $productSW->getConfiguratorSet()->getId() > 0) {  // Varkombi
+            $pk = '';
+        }
+
+        $productSeo = sprintf('%s %s%s',
+            $productSW->getName(),
+            $detailSW->getAdditionalText(),
+            $pk
+        );
+
+        if (strlen($productSeo) > 60) {
+            $pos = strpos($productSeo, ' ', 60);
+            if ($pos === false) {
+                $pos = 60;
+            }
+
+            $productSeo = substr($productSeo, 0, $pos);
+        }
+
+        return $seo->create(
+            sprintf('%s %s',
+                $productSeo,
+                $detailSW->getNumber()
+            )
+        );
+    }
+
+    protected function findParentImage($articleId, $imageFile)
+    {
+        $results = Shopware()->Db()->fetchAssoc(
+            'SELECT m.path, i.id
+             FROM s_articles_img i
+             JOIN s_media m ON m.id = i.media_id
+             WHERE i.articleID = ' . intval($articleId)
+        );
+
+        if (is_array($results) && count($results) > 0) {
+            foreach ($results as $result) {
+                $file = Path::combine(Shopware()->DocPath(), $result['path']);
+                if (file_exists($file)) {
+                    if (md5_file($imageFile) === md5_file($file)) {
+                        return $this->Manager()->find('Shopware\Models\Article\Image', (int) $result['id']);
+                        /*
+                        return $this->Manager()->createQueryBuilder()->select(
+                                'image',
+                                'media'
+                            )
+                            ->from('Shopware\Models\Article\Image', 'image')
+                            ->join('image.media', 'media')
+                            ->where('image.id = :imageId')
+                            ->setParameter('imageId', (int) $result['id'])
+                            ->getQuery()->getResult();
+                        */
+                    }
+                }
+            }
+        }
+
+        return null;
+    }
+
+    protected function buildProductImageMappings(ArticleImageSW $parentSW, $detailId = null)
+    {
+        try {
+            $builder = $this->Manager()->createQueryBuilder()->select(
+                'detail',
+                'options'
+            )
+                ->from('Shopware\Models\Article\Detail', 'detail')
+                ->join('detail.configuratorOptions', 'options')
+                ->join('detail.images', 'images')
+                ->where('images.parent = :parentId')
+                ->setParameter('parentId', $parentSW->getId());
+
+            if ($detailId !== null) {
+                $builder->andWhere('detail.id != :detailId')
+                    ->setParameter('detailId', $detailId);
+            }
+
+            $parentSW = $this->findArticleImage($parentSW->getId());
+            $detailsSW = $builder->getQuery()->getResult();
+
+            foreach ($detailsSW as $detailSW) {
+                $mappingSW = new MappingSW();
+                $mappingSW->setImage($parentSW);
+
+                $rules = [];
+                foreach ($detailSW->getConfiguratorOptions() as $optionSW) {
+                    $ruleSW = new RuleSW();
+                    $ruleSW->setOption($optionSW);
+                    $ruleSW->setMapping($mappingSW);
+                    $this->Manager()->persist($ruleSW);
+                }
+
+                $mappingSW->setRules($rules);
+                $this->Manager()->persist($mappingSW);
+            }
+        } catch (\Exception $e) {
+            Logger::write(ExceptionFormatter::format($e), Logger::ERROR, 'database');
+        }
+    }
+
+    private function deleteProductImageMappings($imageParentId)
+    {
+        try {
+            Shopware()->Db()->query(
+                'DELETE m, r
+                    FROM s_article_img_mappings m
+                    LEFT JOIN s_article_img_mapping_rules r ON r.mapping_id = m.id
+                    WHERE m.image_id = ?',
+                [$imageParentId]
+            );
+        } catch (\Exception $e) {
+            Logger::write(ExceptionFormatter::format($e), Logger::ERROR, 'database');
+        }
     }
 }