+++ src/jtl/Connector/Shopware/Mapper/MeasurementUnit.php
@@ -134,7 +134,7 @@ class MeasurementUnit extends DataMapper
                 $shopMapper = Mmc::getMapper('Shop');
                 $shops = $shopMapper->findByLocale($locale->getLocale());
 
-                if ($shops === null) {
+                if ($shops === null || (is_array($shops) && count($shops) == 0)) {
                     Logger::write(sprintf('Could not find any shop with locale (%s) and iso (%s)', $locale->getLocale(), $iso), Logger::WARNING, 'database');
 
                     continue;